chore: coverts gallery pagintion components to tsx
